requirements.txt
django
django-extensions
django-heroku
dj-database-url
python-dotenv
pydotplus
psycopg2
psycopg2-binary==2.8.6
whitenoise
secret-key-generator
django-multiselectfield
django-money
django-address
django-image-cropping
django-phone-field
django-ckeditor
django-allauth == 0.43.0
pillow
